+++ presentation/src/main/java/com/agileburo/anytype/presentation/desktop/DashboardView.kt
@@ -3,8 +3,11 @@ package com.agileburo.anytype.presentation.desktop
 import com.agileburo.anytype.domain.common.Id
 
 sealed class DashboardView {
+
+    abstract val id: Id
+
     data class Document(
-        val id: Id,
+        override val id: Id,
         val target: Id,
         val title: String? = null,
         val emoji: String? = null,