+++ presentation/src/main/java/com/agileburo/anytype/presentation/mapper/MapperExtension.kt
@@ -188,8 +188,9 @@ fun Block.Content.Text.marks(): List<Markup.Mark> = marks.mapNotNull { mark ->
     }
 }
 
-fun HomeDashboard.toView(builder: UrlBuilder): List<DashboardView.Document> =
-    children.mapNotNull { id ->
+fun HomeDashboard.toView(
+    builder: UrlBuilder
+): List<DashboardView.Document> = children.mapNotNull { id ->
     blocks.find { block -> block.id == id }?.let { model ->
         when (val content = model.content) {
             is Block.Content.Link -> {